notification-service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pika==1.3.2
mysqlclient==2.2.0
python-dotenv==1.0.0
